requirements.txt
# Production Requirements for SchoolPlatform
# ==========================================

# Core Django
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Cache and Background Tasks
django-redis==5.4.0
celery==5.3.6
redis==5.0.1

# CORS and Security
django-cors-headers==4.3.1

# Performance Monitoring
django-debug-toolbar==4.2.0
sentry-sdk==1.43.0

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0
django-filters==23.5

# API Documentation
drf-spectacular==0.27.1

# Content Management
django-modelcluster==6.2.1
django-taggit==4.0.0

# Production Server
gunicorn==21.2.0

# Environment Management
python-decouple==3.8

# File Storage (optional)
# django-storages==1.14.2
# boto3==1.34.51

# Monitoring (optional)
# django-health-check==3.17.0

# Development Tools (remove in production)
# black==24.2.0
# flake8==7.0.0
# isort==5.13.2
